Fix: dequeue one fiber every once in a while (MT)
We previously tried to grab fibers from the global queue which calls
`Runnables#push` that may overflow the local queue and thus try to push
to the global queue, leading to a deadlock.

Replaces the every once in a while dequeue from the global queue with a
single dequeue, which returns one fiber that can be resumed immediately.