Fix: explicitly exit from Crystal.main
When the default execution context is MT, the main thread is hijacked to
become a scheduler of the MT context, yet unlike other threads that will
resume their main fiber to park themselves (it will never be stolen) the
main thread's main fiber is running user code, and as thus will be
enqueued and stolen like any other fiber.

We're already messing the thread stacks with the fiber context switches,
and we're merely doing just that when another thread resumes the main
fiber: continue the fiber.

Still, this patch preventively fixes `Crystal.main` to explicitly call
`LibC._exit` (callable from any thread) instead of returning the exit
status to avoid returning to main, which might expect the main thread
to be the one returning (not another thread) and confuse the OS or the
libc. Let's err on the safe side.

This has been tested to work correctly on Linux: thread #6 resumed the
main fiber of thread #0 and exited properly.