Fix: dead fiber stack recycling (no more segfaults with MT)
I got segfaults whenever I tried to reuse the stacks of dead fibers,
even after the bugfix to only release the stack _after_ the context
switch.

The problem is that after the fiber context switch, we returned to the
`Scheduler#resume` method at the moment the fiber was saved (i.e. local
variables aren't the same before/after the context switch).

Because of work stealing, the fiber may have been suspended by scheduler
A, be stolen, then be resumed by scheduler B, hence run in another
thread, and local variables aren't the same... which includes `self`
that refers to scheduler A despite having been called from scheduler B!

This means that we can't trust anything after a call to `#swapcontext`
or `#resume` in general, we can't trust local or instance variables or
even calling a method on `self`: they would be called on the previous
scheduler (which is wrong)!

NOTE: the run loop fibers are treated in a special way, they're never
enqueued and always suspended and resumed from the scheduler they belong
to, so in that very specific case we can trust the local and instance
variables and it's safe to keep calling methods on `self`!

ALSO FIXES a stack leak by releasing the dead fiber stack in
`Fiber#run` just like we release the GC read lock.