Refactor EC::Isolated to own the thread (really _one_ fiber)
Instead of extending the EC::SingleThreaded, EC::Isolated is now a
distinct implementation with a single thread and a single fiber (the
thread's main fiber), while still being to wait on the event loop or
communicate with other contexts through channels.

Note: unlike other contexts, where threads loop and park themselves,
the thread of isolated contexts will terminate as soon as the block
terminates.