Refactor EC::SingleThreaded to remove the deep sleep fiber
The deep sleep fiber was only listening when both runnables and the
event loop were empty, but we still need to be able to communicate with
the context from an external context (e.g. channel, spawn) and that must
interrupt the event loop.

Since we can now interrupt a blocked loop that can also be made to wait
forever, we don't need the fiber channel anymore, and can just interrupt
the loop.