Fix overload order for `MultiThreaded.new`
This is a workaround for faulty overload ordering by the compiler that cannot be fixed without breaking backwards compatibilty (https://github.com/crystal-lang/crystal/pull/10711).

The compile time flag `-Dpreview_overload_order` can be used to opt-in to the fixed behaviour. But without it, the non-protected def must be after the protected one.