Fix: run loop keeps spinning instead of restarting
The run loop starts to spin when there is nothing to do, which creates
a loop inside the loop, and the `runnable` macro would call `next` to
restart the loop from the beginning... which in that case would only
restart from the spinning part!

The schedulers' run loop was refactored into an outer loop that calls
an inner iteration method that now _returns_ the fiber to be resumed or
nil to skip to the next loop iteration.

This change also enabled to simplify the behavior and to remove the
tricky runnable macro, replaced with an explicit `return fiber if
fiber`; removing the spinning and idle flags is now centralized to the
outer loop.